+++ src/components/editor/editor.scss
@@ -1,0 +1,18 @@
+textarea,
+input {
+    border: 1px solid rgb(192, 192, 192);
+    appearance: none;
+    background: transparent;
+    color: rgb(192, 192, 192);
+}
+
+.name {
+    width: 100%;
+    margin: 1rem 0;
+    font-size: 2rem;
+}
+
+.data {
+    width: 100%;
+    min-height: 500px;
+}
