+++ src/components/editor/editor.svelte
@@ -1,22 +1,5 @@
-<style>
-    textarea,
-    input {
-        border: 1px solid rgb(192, 192, 192);
-        appearance: none;
-        background: transparent;
-        color: rgb(192, 192, 192);
-    }
-
-    .name {
-        width: 100%;
-        margin: 1rem 0;
-        font-size: 2rem;
-    }
-
-    .data {
-        width: 100%;
-        min-height: 500px;
-    }
+<style lang="scss">
+@import "editor";
 </style>
 
 <input type="text" class="name" placeholder="Name..." bind:value="{data.name}">
@@ -27,17 +10,17 @@ import service from "/src/service.js";
 
 const data = {
     name : "",
-    note : ""
+    note : "",
 };
 
 $: service.send({
     type : "plugin:urlContext:CHANGE",
-    data
+    data,
 });
 
 $: ({
-    note = {}
-} = $service.context)
+    note = {},
+} = $service.context);
 
 $: if(!data.name && note.name) {
     data.name = note.name;
@@ -46,5 +29,4 @@ $: if(!data.name && note.name) {
 $: if(!data.note && note.name) {
     data.note = note.note;
 }
-
 </script>
