+++ src/main.js
@@ -3,7 +3,7 @@ import Main from "./main.svelte";
 import "./styles/main.scss";
 
 const main = new Main({
-    target : document.body,
+    target: document.body,
 });
 
 export default main;
